♻️ Refactor: kebab, subFolderUtil에 전달되는 함수 객체의 타입 지정
